+++ src/scripts/services/router.service.ts
@@ -1,58 +1,34 @@
-import { Router } from "./router";
+import { type Route, Router } from "./router";
+import { renderLogin } from "../pages/login";
+import { renderDashboard } from "../pages/dashboard";
+import { renderQuiz } from "../pages/quiz";
+import { renderStatistic } from "../pages/statistic";
 
-let isAuth: boolean = false;
+let isAuth: boolean = true;
 
-function renderLogin() {
-  const root = document.getElementById("app")!;
-  root.innerHTML = `
-    <h1>Login Page</h1>
-    <button id="loginBtn">Login</button>
-  `;
-
-  document.getElementById("loginBtn")!.onclick = () => {
-    isAuth = true;
-    router.navigate("/dashboard");
-  };
-}
-
-function renderQuiz() {
-  const root = document.getElementById("app")!;
-  root.innerHTML = `
-    <h1>Quiz Page</h1>
-    <button id="backBtn">Back</button>
-  `;
-
-  document.getElementById("backBtn")!.onclick = () => {
-    router.navigate("/dashboard");
-  };
-}
-
-function renderDashboard() {
-  const root = document.getElementById("app")!;
-  root.innerHTML = `
-    <h1>Dashboard</h1>
-    <button id="logoutBtn">Logout</button>
-    <button id="quizBtn">Go to Quiz</button>
-  `;
-
-  document.getElementById("logoutBtn")!.onclick = () => {
-    isAuth = false;
-    router.navigate("/");
-  };
-
-  document.getElementById("quizBtn")!.onclick = () => {
-    router.navigate("/quiz");
-  };
-}
-
-const routes = [
-  { path: "/", render: renderLogin },
-  { path: "/dashboard", render: renderDashboard, protected: true },
-  { path: "/quiz", render: renderQuiz, protected: true }
+const routes: Route[] = [
+  {
+    path: "/",
+    render: () => renderLogin(router, (value) => (isAuth = value))
+  },
+  {
+    path: "/dashboard",
+    render: () => renderDashboard(router, (value) => (isAuth = value)),
+    protected: true
+  },
+  {
+    path: "/quiz",
+    render: () => renderQuiz(router),
+    protected: true
+  },
+  {
+    path: "/statistic",
+    render: () => renderStatistic(router),
+    protected: true
+  }
 ];
 
 const router = new Router(routes, () => isAuth);
 router.init();
-
