+++ src/scripts/services/router.ts
@@ -1,4 +1,4 @@
-type Route = {
+export type Route = {
   path: string;
   render: () => void;
   protected?: boolean;
